+++ index.ts
@@ -68,6 +68,8 @@ export class Station {
      *
      * The constant {@linkcode CPAY} is used for contactless-only stations.
      *
+     * The constant {@linkcode SFA} is used for Stratford International (NR).
+     * 
      * @example
      * [1]
      */
@@ -111,19 +113,70 @@ export enum Mode {
 export const stations : Station[] = _stations.map(Station.fromJson);
 
 /**
+ * All TfL lines in the order to be displayed
+ */
+export const tflLines = [
+    "Bakerloo",
+    "Central",
+    "Circle",
+    "District",
+    "Hammersmith & City",
+    "Jubilee",
+    "Metropolitan",
+    "Northern",
+    "Piccadilly",
+    "Victoria",
+    "Waterloo & City",
+    "DLR",
+    "Overground",
+    "Elizabeth line",
+    "Tram"
+];
+
+function getSortIndex(line : string) {
+    const tflIndex = tflLines.indexOf(line);
+    return tflIndex !== -1 ? tflIndex : tflLines.length;
+}
+
+/**
+ * All lines in the order to be displayed
+ */
+export const lines = [...new Set<string>(stations.flatMap(station => station.lines))]
+    .sort((a, b) => {
+        const a_index = getSortIndex(a);
+        const b_index = getSortIndex(b);
+        if (a_index === b_index) {
+            return a === b ? 0 : a.localeCompare(b, 'en-GB');
+        }
+        return a_index - b_index;
+    });
+
+/**
  * A pseudo-zone representing contactless-only stations
  */
 export const CPAY = 16;
 
 /**
- * Get all stations within the specified {@link zones}, {@link modes} and [sides of the river]{@link riverBanks}.
+ * A pseudo-zone with Stratford International (NR) only
  */
-export function getBasket(zones : number[] | undefined, modes : Mode[] | undefined, riverBanks : RiverBank[] | undefined) {
+export const SFA = 0;
+
+/**
+ * Get all stations within the specified {@link zones}, {@link modes} 
+ * and [sides of the Thames]{@link riverBanks} on specified {@link lines}.
+ */
+export function getBasket(
+    zones? : number[],
+    modes? : Mode[],
+    riverBanks? : RiverBank[],
+    lines? : string[]
+) {
     return stations.filter(
         station =>
             match(zones, station.zones)
             && match(modes, station.modes)
             && match(riverBanks, station.riverBanks)
+            && match(lines, station.lines)
     );
 }
 
@@ -145,7 +198,7 @@ export function fromString(id : string) {
 
 /**
  * Get a random set of {@link count} stations from the {@link basket}, optionally
- * starting at {@link startingStation}.
+ * starting at [a specified station]{@link startingStation}.
  */
 export function generate(
     count: number